send modals to screens
