+++ packages/app/components/icons/IconBack.tsx
@@ -1,0 +1,27 @@
+import { ColorTokens } from '@my/ui/types'
+import { themed, IconProps } from '@tamagui/helpers-icon'
+import { memo } from 'react'
+import { Path, Svg } from 'react-native-svg'
+
+const Back = (props: IconProps) => {
+  const { size, color, ...rest } = props
+  return (
+    <Svg
+      viewBox="0 0 24 24"
+      color={color as ColorTokens | undefined}
+      width={size ?? 24}
+      height={size ?? 24}
+      fill={'none'}
+      {...rest}
+    >
+      <Path
+        d="M16.2 19.2002L7.19995 12.0002L16.2 4.8002"
+        stroke="#C3AB8E"
+        strokeWidth="2"
+        strokeLinecap="round"
+      />
+    </Svg>
+  )
+}
+const IconBack = memo(themed(Back))
+export { IconBack }
