+++ packages/app/features/send/components/send/index.tsx
@@ -1,0 +1,2 @@
+export * from './confirm-modal'
+export * from './SendItButton'
