+++ packages/app/features/send/screens/send/index.tsx
@@ -1,2 +1,3 @@
+export * from './send'
 export * from './send-tag'
 export * from './send-it'
